Split `StructureID` enum from `Common.h`
