+++ OPHD/ophd.vcxproj
@@ -268,6 +268,7 @@
     <ClInclude Include="DirectionOffset.h" />
     <ClInclude Include="EnumConnectorDir.h" />
     <ClInclude Include="EnumDirection.h" />
+    <ClInclude Include="EnumStructureID.h" />
     <ClInclude Include="GraphWalker.h" />
     <ClInclude Include="IOHelper.h" />
     <ClInclude Include="Map\MapCoordinate.h" />
